refactor(formatToKRW): 해당 유틸 함수의 경로 수정
